### 2025-05-17 #8244 @editorialbot
[PRE REVIEW]: AOUSDOHtools: An R Package for Social Determinants of Health Survey data in the All of Us Research Program
**Submitting author:** @zhd52 (<a href="http://orcid.org/0009-0004-6658-5205">Zhirui Deng</a>)
**Repository:** https://github.com/zhd52/AOUSDOHtools
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.4
