### 2025-05-17 #8244 @arfon
@zhd52 please see the example paper for how to reference the bibtex file in your header https://joss.readthedocs.io/en/latest/example_paper.html

We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
