### 2025-07-13 #8244 @zhd52
Thanks so much, @smasongarrison ! Really appreciate your offer, hope we can have you on board once an editor is assigned!

@arfon just flagging this in case it's helpful once the paper moves off the waitlist. Thanks!
