### 2025-08-26 #8244 @goldingn
Hi @zhd52 can you please try to fix the `paper.md` file? Here's a reference file: https://joss.readthedocs.io/en/latest/example_paper.html

I can see that you are missing the date and bibliography fields in the header, which is probably what is causing the compilation error.
