### 2025-08-31 17fa2f2
Fix JOSS paper header: add date and bibliography, 20250831
