### 2025-09-24 #8244 @goldingn
Hi @philerooski, would you be interested in reviewing this R package submission for JOSS?
There are some similarities (health data, wearables) to your 2020 JOSS paper
