### 2025-10-16 #8244 @goldingn
Thanks @zhd52 for the prod and sorry for the delay. I'd gratefully take any suggestions from you for potential reviewers who work in this space. We only need one more.

Quoting from the top of the thread:
> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
