### 2025-10-24 #8244 @goldingn
Hi @hantswilliams, would you be interested in reviewing this submission for JOSS?
