### 2025-11-06 7dbaa1f
table merging to readme
addition of dependencies for table merging
