### 2025-11-06 ccd6804
Refine installation steps in README.md
Updated installation instructions for AOUSDOHtools.
