### 2025-11-07 #9306 @smasongarrison
Other notes: Love the Function Map
