### 2025-11-07 dc8d70a
fix references
