+++ paper.md
@@ -40,7 +40,8 @@ journal: JOSS
 
 The package is designed for use within the *All of Us* Researcher Workbench, a secure cloud-based platform where the de-identified data are accessed and analyzed. The package is compatible with both Jupyter and RStudio environments hosted on the platform. `AOUSDOHtools` automates the data cleaning, recoding, scoring, and variable construction, which enables researchers to generate interpretable SDOH scores for downstream analysis.
 
-The package is openly developed and maintained on GitHub and available through CRAN (cran.r-project.org/package=AOUSDOHtools) and GitHub (github.com/zhd52/AOUSDOHtools). It is intended to facilitate equitable and scalable research by making complex SDOH survey data accessible and analysis-ready for approved researchers working within the *All of Us* ecosystem.
+The package is openly developed and maintained on GitHub and available through CRAN ([cran.r-project.org/package=AOUSDOHtools](https://cran.r-project.org/package=AOUSDOHtools)) and GitHub ([github.com/zhd52/AOUSDOHtools](https://github.com/zhd52/AOUSDOHtools)). It is intended to facilitate equitable and scalable research by making complex SDOH survey data accessible and analysis-ready for approved researchers working within the *All of Us* ecosystem.
+
 
 -   **Figure 0:** `AOUSDOHtools` hex sticker. Created with R package `hexSticker` [@hexSticker2020].
 
@@ -48,21 +49,22 @@ The package is openly developed and maintained on GitHub and available through C
 
 ## ***All of Us (AOU)***
 
-The *All of Us* Research Program (allofus.nih.gov), led by the National Institutes of Health, aims to create a large and diverse health database by enrolling over one million participants across the United States. The program supports research focused on individualized prevention, diagnosis, and treatment [@AllOfUsNIH2025].
+The *All of Us* Research Program ([allofus.nih.gov](https://www.allofus.nih.gov)), led by the National Institutes of Health, aims to create a large and diverse health database by enrolling over one million participants across the United States. The program supports research focused on individualized prevention, diagnosis, and treatment [@AllOfUsNIH2025].
 
-Participants provide data through self-reported surveys, electronic health records (EHR), physical measurements, wearable devices (e.g. FitBit), and biospecimen collection (e.g. urine and blood specimens), made available to approved researchers via the secure cloud-based Researcher Workbench (researchallofus.org) [@AllOfUsNEJM2019]. A major strength of *All of Us* is its explicit focus on health equity, particularly through the inclusion of historically underrepresented populations in biomedical research [@AllOfUsNIH2025].
+Participants provide data through self-reported surveys, electronic health records (EHR), physical measurements, wearable devices (e.g. FitBit), and biospecimen collection (e.g. urine and blood specimens), made available to approved researchers via the secure cloud-based Researcher Workbench ([researchallofus.org](https://www.researchallofus.org)) [@AllOfUsNEJM2019]. A major strength of *All of Us* is its explicit focus on health equity, particularly through the inclusion of historically underrepresented populations in biomedical research [@AllOfUsNIH2025].
+
 
 Despite these strengths, the complexity of the available data, especially the survey components, can present analytic challenges. Standardized and scalable tools are needed to support consistent data processing and analysis, but such tools are not included in the platform by default [@Grayson2022].
 
 ## **Social Determinants of Health (SDOH) Survey**
 
[…]
