### 2025-11-10 706ea2e
Revise JOSS paper and README per reviewer feedback, 20251110
