+++ paper.md
@@ -42,8 +42,7 @@ The package is designed for use within the *All of Us* Researcher Workbench, a s
 
 The package is openly developed and maintained on GitHub and available through CRAN ([cran.r-project.org/package=AOUSDOHtools](https://cran.r-project.org/package=AOUSDOHtools)) and GitHub ([github.com/zhd52/AOUSDOHtools](https://github.com/zhd52/AOUSDOHtools)). It is intended to facilitate equitable and scalable research by making complex SDOH survey data accessible and analysis-ready for approved researchers working within the *All of Us* ecosystem.
 
-
--   **Figure 0:** `AOUSDOHtools` hex sticker. Created with R package `hexSticker` [@hexSticker2020].
+![**Figure 0.** `AOUSDOHtools` hex sticker. Created with R package `hexSticker` [@hexSticker2020].](figures/Hex_Sticker.jpg){width="258"}
 
 # **Statement of Need**
 
@@ -52,7 +51,6 @@ The package is openly developed and maintained on GitHub and available through C
 The *All of Us* Research Program ([allofus.nih.gov](https://www.allofus.nih.gov)), led by the National Institutes of Health, aims to create a large and diverse health database by enrolling over one million participants across the United States. The program supports research focused on individualized prevention, diagnosis, and treatment [@AllOfUsNIH2025].
 
 Participants provide data through self-reported surveys, electronic health records (EHR), physical measurements, wearable devices (e.g. FitBit), and biospecimen collection (e.g. urine and blood specimens), made available to approved researchers via the secure cloud-based Researcher Workbench ([researchallofus.org](https://www.researchallofus.org)) [@AllOfUsNEJM2019]. A major strength of *All of Us* is its explicit focus on health equity, particularly through the inclusion of historically underrepresented populations in biomedical research [@AllOfUsNIH2025].
-
 
 Despite these strengths, the complexity of the available data, especially the survey components, can present analytic challenges. Standardized and scalable tools are needed to support consistent data processing and analysis, but such tools are not included in the platform by default [@Grayson2022].
 
@@ -64,70 +62,25 @@ Transforming survey responses into meaningful constructs for research, such as s
 
 ## **Scoring Social Determinants of Health (SDOH) Constructs**
 
-The `AOUSDOHtools` R Package (@Deng2025; [cran.r-project.org/package=AOUSDOHtools](https://cran.r-project.org/package=AOUSDOHtools)) was developed to address this gap in available tools by implementing literature-informed scoring logic for 14 well-defined SDOH constructs (***Figure 1***). `AOUSDOHtools` builds upon a user guide developed by @Koleck2024 and includes constructs such as Neighborhood Cohesion, Social Support, and Perceived Stress, each assessed using standardized instruments. Please see @Koleck2024 for definitions and descriptions of the SDOH constructs and source instrument information. A total of 30 functions for scoring are available. Six of the constructs have one scoring option. Eight of the constructs have multiple scoring options. `AOUSDOHtools` is intended for use exclusively within the *All of Us* Researcher Workbench [@AllOfUsWorkbench2025] (Jupyter or Rstudio), respecting privacy and data governance regulations.
[…]
