+++ paper.bib
@@ -13,19 +13,19 @@
   title        = {All of Us Research Program},
   author       = {{National Institutes of Health}},
   year         = {2025},
-  howpublished = {Retrieved from https://allofus.nih.gov/}
+  url          = {https://allofus.nih.gov/}
 }
 
 @Misc{AllOfUsWorkbench2025,
   title        = {All of Us Researcher Workbench},
   author       = {{All of Us Research Hub}},
   year         = {2025},
-  howpublished = {Retrieved from https://www.researchallofus.org/}
+  url          = {https://www.researchallofus.org/}
 }
 
-@Manual{Deng2025,
+@Manual{AOUSDOHtools2025,
   title        = {AOUSDOHtools: Analyzing AOU SDOH Survey Data},
-  author       = {Deng, Z. and Koleck, T. A. and Dreisbach, C. and Zhang, C. and Higgins, P. D. R.},
+  author       = {Deng, Z. and Koleck, T. A. and Zhang, C. and Higgins, P. D. R. and Dreisbach, C.},
   year         = {2025},
   note         = {R package version 1.0.4},
   doi          = {10.32614/cran.package.aousdohtools}
@@ -36,7 +36,7 @@
   author       = {Wickham, H. and Hester, J. and Chang, W. and Bryan, J.},
   year         = {2022},
   note         = {R package version 2.4.5},
-  url          = {https://CRAN.R-project.org/package=devtools}
+  doi          = {10.32614/cran.package.devtools}
 }
 
 @Manual{DiagrammeR2024,
@@ -44,7 +44,7 @@
   author       = {Iannone, R. and Roy, O.},
   year         = {2024},
   note         = {R package version 1.0.11},
-  url          = {https://CRAN.R-project.org/package=DiagrammeR}
+  doi          = {10.32614/cran.package.diagrammer}
 }
 
 @Manual{DiagrammeRsvg2016,
@@ -52,7 +52,7 @@
   author       = {Iannone, R.},
   year         = {2016},
   note         = {R package version 0.1},
-  url          = {https://CRAN.R-project.org/package=DiagrammeRsvg}
+  doi          = {10.32614/cran.package.diagrammersvg}
 }
 
 @Manual{dplyr2023,
@@ -60,7 +60,7 @@
   author       = {Wickham, H. and François, R. and Henry, L. and Muller, K. and Vaughan, D.},
   year         = {2023},
   note         = {R package version 1.1.4},
-  url          = {https://CRAN.R-project.org/package=dplyr}
+  doi          = {10.32614/cran.package.dplyr}
 }
 
 @Article{Grayson2022,
@@ -79,7 +79,7 @@
   author       = {Yu, G. and Gatto, L. and Rainer, J. and Gibb, S. and Corty, R. and Neitmann, T.},
   year         = {2020},
   note         = {R package version 0.4.9},
-  url          = {https://CRAN.R-project.org/package=hexSticker}
+  doi          = {10.32614/cran.package.hexsticker}
 }
 
 @Article{Koleck2024,
@@ -97,7 +97,7 @@
   title        = {Social determinants of health. Healthy People 2030},
   author       = {{Office of Disease Prevention and Health Promotion}},
   year         = {2025},
-  howpublished = {Retrieved from https://odphp.health.gov/healthypeople/priority-areas/social-determinants-health}
+  url          = {https://odphp.health.gov/healthypeople/priority-areas/social-determinants-health}
 }
 
 @Manual{purrr2025,
@@ -105,7 +105,7 @@
   author       = {Wickham, H. and Henry, L.},
   year         = {2025},
   note         = {R package version 1.0.4},
-  url          = {https://CRAN.R-project.org/package=purrr}
[…]
