### 2025-11-10 #9306 @zhd52
@hantswilliams @smasongarrison @goldingn I've revised the paper and related files based on your feedback, and all updates are now in the original repository.

I’ve merged the pull requests, fixed missing DOIs and references, added contribution guidelines, clarified how the package compares to others, and noted All of Us data privacy considerations and Jupyter Notebook usage. I also moved the installation and example code into vignettes, addressed missing data handling, and added figures to the paper.

Thank you all for your time and thoughtful reviews! Please let me know if anything else is needed. I’m looking forward to the next steps!
