### 2025-11-10 a4c5c05
Minor edits, 20251110
