### 2025-11-10 #9306 @smasongarrison
@zhd52 Thanks. I've marked off some of the feedback. But, I'd really like to see some contextualizing of other specific software. Surely, there exist other specific tools like this for other datasets.

> "While other R packages offer general scoring utilities, `AOUSDOHtools` uniquely integrates the *All of Us* variable naming conventions, metadata harmonization, and privacy-preserving workflows."

It's not so much about saying what makes this software unique. It's about helping readers finding other relevant software that has tried to tackle similar challenges.

I also think your readers would benefit if you briefly explained what is in @Koleck2024, rather than just adding the citation.
