### 2025-11-14 a7f596f
Revise JOSS paper per reviewer feedback, 20251114
