+++ paper.bib
@@ -100,6 +100,22 @@
   url          = {https://odphp.health.gov/healthypeople/priority-areas/social-determinants-health}
 }
 
+@Manual{PROscorer2023,
+  title        = {PROscorer: Functions to Score Commonly-Used Patient-Reported Outcome (PRO) Measures and Other Psychometric Instruments},
+  author       = {Baser, R.},
+  year         = {2023},
+  note         = {R package version 0.0.4},
+  doi          = {10.32614/cran.package.proscorer}
+}
+
+@Manual{psych2025,
+  title        = {psych: Procedures for Psychological, Psychometric, and Personality Research},
+  author       = {Revelle, W.},
+  year         = {2025},
+  note         = {R package version 2.5.6},
+  doi          = {10.32614/cran.package.psych}
+}
+
 @Manual{purrr2025,
   title        = {purrr: Functional Programming Tools},
   author       = {Wickham, H. and Henry, L.},
@@ -122,6 +138,14 @@
   year         = {2025},
   note         = {R package version 2.6.2},
   doi          = {10.32614/cran.package.rsvg}
+}
+
+@Manual{scorecard2025,
+  title        = {scorecard: Credit Risk Scorecard},
+  author       = {Xie, S.},
+  year         = {2025},
+  note         = {R package version 0.4.5},
+  doi          = {10.32614/cran.package.scorecard}
 }
 
 @Misc{SurveyExplorer2025,
