### 2025-11-17 #9306 @smasongarrison
@zhd52 @goldingn I think I've got everything checked off my list. I do think that there's probably a nicer way to wrap text around figure 2, but that might be beyond editorialbot's abilities.
