### 2025-11-21 #9306 @goldingn
Fantastic, thanks @hantswilliams!

@zhd52 now that the reviewers are happy I'll move onto my checklist to set up the archive and assess formatting etc. Once that is done I can recommend acceptance and a track editor will take over the final steps.

As per the **instructions in the message above**, I need you to: check the author information and licence, make and archive a release, and leave me a message here telling me the release number and the DOI.
