### 2025-11-24 #9306 @zhd52
@hantswilliams Thank you so much for reviewing this work!

@goldingn Thanks for the instructions. I've completed these post-review tasks.

- Software release version: v1.1.0
- Zenodo archive DOI: 10.5281/zenodo.17693884

Please let me know if anything else is needed. Thank you!
