+++ paper.md
@@ -62,7 +62,7 @@ Transforming survey responses into meaningful constructs for research, such as s
 
 ## **Scoring Social Determinants of Health (SDOH) Constructs**
 
-The `AOUSDOHtools` R Package (@AOUSDOHtools2025; [cran.r-project.org/package=AOUSDOHtools](https://cran.r-project.org/package=AOUSDOHtools)) was developed to address the absence of standardized tools for processing and scoring the *All of Us* SDOH Survey. It operationalizes literature-informed scoring logic for 14 SDOH constructs (***Figure 1***), such as Neighborhood Cohesion, Social Support, and Perceived Stress. Each construct includes well-defined item groupings, reverse-coding procedures, and aggregation rules defined in the *All of Us* SDOH User Guide [@Koleck2024]. The package includes 30 functions that automate scoring across these constructs, enabling reproducible and scalable analyses within the *All of Us* Researcher Workbench [@AllOfUsWorkbench2025]. Six of the constructs have one scoring option. Eight of the constructs have multiple scoring options. `AOUSDOHtools` is intended for use exclusively within the *All of Us* Researcher Workbench (Jupyter or Rstudio), respecting privacy and data governance regulations.
+The `AOUSDOHtools` R Package (@AOUSDOHtools2025; [cran.r-project.org/package=AOUSDOHtools](https://cran.r-project.org/package=AOUSDOHtools)) was developed to address the absence of standardized tools for processing and scoring the *All of Us* SDOH Survey. It operationalizes literature-informed scoring logic for 14 SDOH constructs (***Figure 2***), such as Neighborhood Cohesion, Social Support, and Perceived Stress. Each construct includes well-defined item groupings, reverse-coding procedures, and aggregation rules defined in the *All of Us* SDOH User Guide [@Koleck2024]. The package includes 30 functions that automate scoring across these constructs, enabling reproducible and scalable analyses within the *All of Us* Researcher Workbench [@AllOfUsWorkbench2025]. Six of the constructs have one scoring option. Eight of the constructs have multiple scoring options. `AOUSDOHtools` is intended for use exclusively within the *All of Us* Researcher Workbench (Jupyter or Rstudio), respecting privacy and data governance regulations.
 
 Several R packages provide general scoring or psychometric utilities, such as `psych` [@psych2025], `PROscorer` [@PROscorer2023], and `scorecard` [@scorecard2025], but these are designed for use across a wide range of datasets and do not implement the unique structure of *All of Us* data. In contrast, `AOUSDOHtools` incorporates the *All of Us* variable naming conventions, metadata alignment, and privacy-preserving workflows required by the controlled Workbench environment.
 
@@ -72,13 +72,13 @@ This package builds directly upon the *All of Us* SDOH User Guide [@Koleck2024],
 
 # **Installation**
 
-The `AOUSDOHtools` package is available on CRAN ([cran.r-project.org/package=AOUSDOHtools](https://cran.r-project.org/package=AOUSDOHtools)) and GitHub ([github.com/zhd52/AOUSDOHtools](https://github.com/zhd52/AOUSDOHtools)). Users can install the stable release from CRAN or the development version from GitHub using the following commands in R [@Rcore2025; @devtools2022].
+The `AOUSDOHtools` package is available on CRAN ([cran.r-project.org/package=AOUSDOHtools](https://cran.r-project.org/package=AOUSDOHtools)) and GitHub ([github.com/zhd52/AOUSDOHtools](https://github.com/zhd52/AOUSDOHtools)). Users can install the stable release from CRAN or the development version from GitHub. Detailed installation instructions are provided in the package README and vignettes [@Rcore2025; @devtools2022].
 
[…]
