### 2025-11-24 #9306 @zhd52
@goldingn I've updated the archive title and made the two requested edits. I noted that the installation code is available in the README and vignettes.

- Software release version: v1.1.1
- Zenodo archive DOI: 10.5281/zenodo.17693884

Thanks so much!
