### 2025-11-28 #9306 @goldingn
Thanks @zhd52! Over to the track editor(s) now
