### 2025-12-01 1e5557b
Fix affiliations, typo, image size, 20251201
