### 2025-12-01 0550a14
Fix image size, 20251201
