### 2025-12-02 4719327
Add JOSS publication and badges, 20251202
